Validate bits argument of FindSmallCoprimeInverse
If the number of bits requested is greater than or equal to the
number of bits of the group's prime, the function will panic, to let
the caller know that a mistake was made.

Added a unit test to test this panic.

Also, added TODO in ModInverse of int.go.
In the future, this function should panic if nil is returned
from the math/big ModInverse function.