Refactor cyclic package to use large.Int or cyclic.Int where appropriate
In order to create a group, large.Ints will be used. They can also
be used in random and shuffle, since the operations in these structures
do not occur under a group.

Group will operate only on cyclicInts, which in turn can only be created
using the constructors provided by the group.

Any attempt to use a cyclicInt created on a different group will cause
a panic on any function under group.go