Make group generator private
Group generator was defined as G, which was being exported, meaning
it could be easily modified by anyone creating a group.

Change it to gen so that it remains private and added a getter function
returning a copy of the value.

Also added getter functions for P, G, PSub1 and PSub1Factor that return
the value as a large.Int or as a cyclic.Int.