Modify InGroup to handle cases when sample is filled with 0 bytes
When len(sample) was > 1 and filled with 0 bytes, InGroup would report true. This change modifies InGroup to revert back to something more like 12055de but refactored for additional clarity in addition to the comparison fix. Tests for the fixed case are included in the parent commit.

h/t to Alex Dupre from the Elixxir Discord for reporting this bug.